requirements.txt
tensorflow>=2.8.0
dvclive
pyyaml